# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "使用gin+vue进行开发的博客平台",
        "title": "Gin-Vue-Blog Swagger API接口文档",
        "contact": {},
        "version": "v1.0.0"
    },
    "host": "127.0.0.1:8888",
    "basePath": "/",
    "paths": {
...
        "/authrequired/admin/admin/new": {
            "post": {
                "description": "注册管理员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "注册管理员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户注册信息",
                        "name": "registerInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/authrequired/admin/deleteAdmin": {
            "delete": {
                "description": "删除指定管理员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "删除指定管理员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "需要删除的管理员ID",
                        "name": "DeleteUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserIDOnlyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/authrequired/admin/deleteUser": {
            "delete": {
                "description": "删除指定用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "删除指定用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "需要删除的用户ID",
                        "name": "DeleteUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserIDOnlyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/authrequired/admin/getAllUsers": {
            "get": {
                "description": "获取所有用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "获取所有用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-array_model_User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/authrequired/admin/modifyAdmin": {
            "patch": {
                "description": "修改指定管理员信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "修改指定管理员信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "要修改的管理员的信息",
                        "name": "ModifyUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ModifyUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/authrequired/admin/modifyAdminPassword": {
            "patch": {
                "description": "修改指定管理员密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "修改指定管理员密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "id,新密码",
                        "name": "ModifyPasswordRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ModifyPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/authrequired/admin/modifyUser": {
            "patch": {
                "description": "修改指定用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "修改指定用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "要修改的用户的信息",
                        "name": "ModifyUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ModifyUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/authrequired/admin/modifyUserPassword": {
            "patch": {
                "description": "修改指定用户密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "修改指定用户密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "id,新密码",
                        "name": "ModifyPasswordRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ModifyPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/authrequired/admin/paginatedUsers": {
            "get": {
                "description": "分页获取用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "分页获取用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键字",
                        "name": "key",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序规则",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-array_model_User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ctype.ImageStoreType": {
            "type": "integer",
            "enum": [
                1,
                2
            ],
            "x-enum-comments": {
                "Local": "存储在本地图片",
                "QiNiuYun": "存储在七牛云的图片"
            },
            "x-enum-varnames": [
                "Local",
                "QiNiuYun"
            ]
        },
        "errhandle.ErrCode": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15
            ],
            "x-enum-varnames": [
                "NoError",
                "InnerError",
                "TokenError",
                "ParamsError",
                "UserExists",
                "UserNonExists",
                "PasswordInvalid",
                "PermissionDenied",
                "SexError",
                "NameExists",
                "EmailFormatError",
                "PasswordTooShort",
                "FileSizeTooLarge",
                "FileTypeNoMatch",
                "VerifyCodeError",
                "OtherError"
            ]
        },
        "model.Article": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "comment_count": {
                    "description": "评论数",
                    "type": "integer"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Comment"
                    }
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "description": "文章的封面",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Image"
                        }
                    ]
                },
                "image_id": {
                    "type": "integer"
                },
                "like_count": {
                    "description": "点赞数",
                    "type": "integer"
                },
                "look_count": {
                    "description": "浏览量",
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.Comment": {
            "type": "object",
            "properties": {
                "article_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.CommonResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/errhandle.ErrCode"
                }
            }
        },
...
        "model.CommonResponse-array_model_User": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                },
                "msg": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/errhandle.ErrCode"
                }
            }
        },
...
        "model.Image": {
            "type": "object",
            "properties": {
                "hash": {
                    "description": "判断图片是否重复",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_store_type": {
                    "$ref": "#/definitions/ctype.ImageStoreType"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.ModifyPasswordRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "new_pwb": {
                    "type": "string"
                },
                "old_pwd": {
                    "type": "string"
                }
            }
        },
        "model.ModifyUserRequest": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                }
            }
        },
        "model.RegisterUserRequest": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                }
            }
        },
        "model.Tag": {
            "type": "object",
            "properties": {
                "articles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Article"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "articles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Article"
                    }
                },
                "avatar": {
                    "description": "头像地址",
                    "type": "string"
                },
                "collect_articles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Article"
                    }
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "description": "权限(管理员，普通用户)",
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                }
            }
        },
        "model.UserIDOnlyRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
    }
}